+++ tests/testthat/test-derive_var_disposition_status.R
@@ -3,7 +3,8 @@ dm <- tibble::tribble(
   "TEST01", "PAT01",
   "TEST01", "PAT02",
   "TEST01", "PAT03",
-  "TEST01", "PAT04"
+  "TEST01", "PAT04",
+  "TEST01", "PAT05"
 )
 
 ds <- tibble::tribble(
@@ -21,7 +22,8 @@ ds <- tibble::tribble(
   "TEST01", "PAT03", "DISPOSITION EVENT", "PROGRESSIVE DISEASE", "2021-05-01",
   "TEST01", "PAT03", "OTHER EVENT", "DEATH", "2022-04",
   "TEST01", "PAT04", "PROTOCOL MILESTONE", "INFORMED CONSENT OBTAINED", "2021-04-02",
-  "TEST01", "PAT04", "PROTOCOL MILESTONE", "RANDOMIZATION", "2021-04-11"
+  "TEST01", "PAT04", "PROTOCOL MILESTONE", "RANDOMIZATION", "2021-04-11",
+  "TEST01", "PAT05", "DISPOSITION EVENT", "SCREEN FAILURE", "2021-03-01"
 )
 
 
@@ -32,7 +34,8 @@ test_that("Derive EOSSTT using default mapping", {
     "TEST01", "PAT01", "DISCONTINUED",
     "TEST01", "PAT02", "COMPLETED",
     "TEST01", "PAT03", "DISCONTINUED",
-    "TEST01", "PAT04", "ONGOING"
+    "TEST01", "PAT04", "ONGOING",
+    "TEST01", "PAT05", "NOT STARTED"
   )
 
   actual_output <- derive_var_disposition_status(
@@ -53,9 +56,10 @@ test_that("Derive EOSSTT using default mapping", {
 test_that("Derive EOTSTT using a study specific mapping", {
   format_eosstt <- function(x) {
     case_when(
+      x == "SCREEN FAILURE" ~ "NOT STARTED",
       x == "COMPLETED" ~ "COMPLETED",
       x == "ADVERSE EVENT" ~ "DISCONTINUED DUE TO AE",
-      x %notin% c("ADVERSE EVENT", "COMPLETED") & !is.na(x) ~ "DISCONTINUED NOT DUE TO AE",
+      x %notin% c("ADVERSE EVENT", "COMPLETED", "SCREEN FAILURE") & !is.na(x) ~ "DISCONTINUED NOT DUE TO AE",
       TRUE ~ "ONGOING"
     )
   }
@@ -64,7 +68,8 @@ test_that("Derive EOTSTT using a study specific mapping", {
     "TEST01", "PAT01", "DISCONTINUED DUE TO AE",
     "TEST01", "PAT02", "COMPLETED",
     "TEST01", "PAT03", "DISCONTINUED NOT DUE TO AE",
-    "TEST01", "PAT04", "ONGOING"
+    "TEST01", "PAT04", "ONGOING",
+    "TEST01", "PAT05", "NOT STARTED"
   )
 
   actual_output <- derive_var_disposition_status(